Refactor to VAFChartUtils.ts and add skeleton for unit tests
